third_party/WebKit/LayoutTests/external/wpt/css/css-ui/outline-018-expected.txt

This is a testharness.js-based test.
FAIL outline-style is animated as a discrete type assert_equals: expected "dotted" but got "solid"
Harness: the test ran to completion.
